+++ .idea/csv-plugin.xml
@@ -10,6 +10,13 @@
             </Attribute>
           </value>
         </entry>
+        <entry key="/reLearning/Inremediate_python/files_directories/csv_file_handling/2018_Central_Park_Squirrel_Census_-_Squirrel_Data.csv">
+          <value>
+            <Attribute>
+              <option name="separator" value="," />
+            </Attribute>
+          </value>
+        </entry>
         <entry key="/reLearning/Inremediate_python/nato_phonetic_alphabet.csv">
           <value>
             <Attribute>
